requirements.txt
pandas
numpy
geopandas
shapely
pyproj
fiona

sqlalchemy
psycopg2-binary
geoalchemy2

python-dotenv

matplotlib
seaborn

jupyterlab

requests
tqdm

folium